requirements.txt
coverage==5.5
fastapi==0.100.1
starlette==0.27.0
uvicorn==0.23.1
freezegun==1.1.0
jaeger-client>=4.1.0
opentracing==2.3.0
opentracing-instrumentation==3.2.1
pydantic==2.1.1
pylint==2.10.2
python-dotenv==0.12.0
jsonmerge==1.7.0
pympler==1.0.1
pytz==2021.3
redis==2.10.6
kafka-python==2.0.1
prometheus-fastapi-instrumentator==6.1.0
requests==2.28.2
pandas==2.2.2
passlib==1.7.4
cryptography==42.0.2
bcrypt==3.2.2
python-jose==3.3.0

# test requirements
pytest==6.2.5
pytest-env==0.6.2
pytest-cov==2.12.1
pytest-asyncio==0.14.0


